fix(runner): keep PATH helpers out of ssh client
